+++ dev_build.sh
@@ -1,0 +1,15 @@
+VERSION_FILE="VERSION"
+source VERSION
+
+# Update version path
+rm -R build/
+python3 setup.py build sdist bdist_wheel
+
+# cd django_dev
+# python mk_docs.py
+
+git add --all
+git commit -m "Building a new version ${VERSION}"
+git tag -a ${VERSION} -m "Building a new version ${VERSION}"
+git push
+git push origin ${VERSION}